requirements.txt
deepspeed==0.14.5
accelerate==0.34.0
datasets==2.20.0
flash-attn==2.6.3
huggingface-hub==0.30.1
numpy==1.26.4
peft==0.11.1
pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.40.2
wandb==0.17.7
pyserini==0.44.0
faiss-gpu==1.7.2